pyproject.toml
[project]
name = "google-pollen-exporter"
version = "0.1.0"
description = "Prometheus(OpenMetrics) exporter for Google Pollen API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dacite>=1.9.2",
    "prometheus-client>=0.21.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
